+++ src/Adapter/Api/StudisSelfserviceFrontendApi.mjs
@@ -1,4 +1,3 @@
-import { METHOD_POST } from "../../Libs/flux-http-api/src/Adapter/Method/METHOD.mjs";
 import { COLOR_SCHEME_DARK, COLOR_SCHEME_LIGHT } from "../../Libs/flux-color-scheme-api/src/Adapter/ColorScheme/COLOR_SCHEME.mjs";
 import { PAGE_CHOICE_SUBJECT, PAGE_COMPLETED, PAGE_CREATE, PAGE_IDENTIFICATION_NUMBER, PAGE_INTENDED_DEGREE_PROGRAM, PAGE_INTENDED_DEGREE_PROGRAM_2, PAGE_LEGAL, PAGE_PERSONAL_DATA, PAGE_PORTRAIT, PAGE_PREVIOUS_STUDIES, PAGE_RESUME, PAGE_START, PAGE_UNIVERSITY_ENTRANCE_QUALIFICATION } from "../Page/PAGE.mjs";
 import { SETTINGS_CACHE_IMPLEMENTATION_CACHE_NAME, SETTINGS_INDEXEDDB_IMPLEMENTATION_DATABASE_NAME, SETTINGS_INDEXEDDB_IMPLEMENTATION_STORE_NAME, SETTINGS_STORAGE_IMPLEMENTATION_KEY_PREFIX } from "../Settings/SETTINGS_IMPLEMENTATION.mjs";
@@ -36,10 +35,9 @@ import { SETTINGS_CACHE_IMPLEMENTATION_CACHE_NAME, SETTINGS_INDEXEDDB_IMPLEMENTA
 /** @typedef {import("../Portrait/PortraitElement.mjs").PortraitElement} PortraitElement */
 /** @typedef {import("../PreviousStudies/PreviousStudies.mjs").PreviousStudies} PreviousStudies */
 /** @typedef {import("../PreviousStudies/PreviousStudiesElement.mjs").PreviousStudiesElement} PreviousStudiesElement */
-/** @typedef {import("../Post/Post.mjs").Post} Post */
 /** @typedef {import("../Post/postFunction.mjs").postFunction} postFunction */
-/** @typedef {import("../Post/PostResult.mjs").PostResult} PostResult */
 /** @typedef {import("../../Libs/flux-pwa-api/src/Adapter/Api/PwaApi.mjs").PwaApi} PwaApi */
+/** @typedef {import("../../Service/Request/Port/RequestService.mjs").RequestService} RequestService */
 /** @typedef {import("../../Libs/flux-localization-api/src/Adapter/SelectLanguage/SelectLanguageElement.mjs").SelectLanguageElement} SelectLanguageElement */
 /** @typedef {import("../../Libs/flux-settings-api/src/Adapter/Api/SettingsApi.mjs").SettingsApi} SettingsApi */
 /** @typedef {import("../Start/Start.mjs").Start} Start */
@@ -107,6 +105,10 @@ export class StudisSelfserviceFrontendApi {
      */
     #pwa_api = null;
     /**
+     * @type {RequestService | null}
+     */
+    #request_service = null;
+    /**
      * @type {SettingsApi | null}
      */
     #settings_api = null;
@@ -134,11 +136,6 @@ export class StudisSelfserviceFrontendApi {
         await this.#getLoadingApi();
         const localization_api = await this.#getLocalizationApi();
         await this.#getPwaApi();
-
-        await css_api.importCss(
-            `${__dirname.substring(0, __dirname.lastIndexOf("/"))}/FormInvalid/FormInvalidElement.css`
-        );
-        await import("../FormInvalid/FormInvalidElement.mjs");
 
         await css_api.importCss(
             `${__dirname}/../Layout/style.css`
@@ -208,77 +205,6 @@ export class StudisSelfserviceFrontendApi {
         this.showFrontend(
             true
         );
-    }
-
-    /**
-     * @returns {Promise<PostResult>}
-     */
-    async #back() {
-        try {
-            await (await this.#getHttpApi()).fetch({
-                url: `${__dirname}/../../api/back`,
-                method: METHOD_POST
-            });
-
-            return {
-                ok: true
-            };
-        } catch (error) {
-            console.error(error);
-
-            const localization_api = await this.#getLocalizationApi();
-
-            return {
-                ok: false,
-                "error-messages": [
-                    Object.fromEntries(await Promise.all(Object.entries((await localization_api.getLanguages()).all).map(async ([
-                        language
-                    ]) => [
-                            language,
-                            await localization_api.translate(
-                                error instanceof Response ? "Server error!" : "Network error!",
-                                null,
-                                null,
-                                language
-                            )
-                        ]
-                    )))
-                ]
-            };
-        }
-    }
-
-    /**
-     * @returns {Promise<GetResult | PostResult>}
-     */
-    async #get() {
-        try {
-            return await (await this.#getHttpApi()).fetch({
-                url: `${__dirname}/../../api/get`
-            });
-        } catch (error) {
-            console.error(error);
-
-            const localization_api = await this.#getLocalizationApi();
-
-            return {
-                ok: false,
-                "error-messages": [
-                    Object.fromEntries(await Promise.all(Object.entries((await localization_api.getLanguages()).all).map(async ([
-                        language
-                    ]) => [
-                            language,
-                            await localization_api.translate(
-                                error instanceof Response ? "Server error!" : "Network error!",
-                                null,
-                                null,
-                                language
-                            )
-                        ]
-                    )))
-                ]
-            };
-        }
     }
 
     /**
@@ -774,6 +700,18 @@ export class StudisSelfserviceFrontendApi {
     }
 
     /**
+     * @returns {Promise<RequestService>}
+     */
+    async #getRequestService() {
+        this.#request_service ??= (await import("../../Service/Request/Port/RequestService.mjs")).RequestService.new(
+            await this.#getHttpApi(),
+            await this.#getLocalizationApi()
+        );
+
+        return this.#request_service;
+    }
+
+    /**
      * @returns {Promise<SettingsApi>}
      */
     async #getSettingsApi() {
@@ -841,44 +779,6 @@ export class StudisSelfserviceFrontendApi {
     }
 
     /**
-     * @returns {Promise<void>}
-     */
-    async #logout() {
-        try {
-            await (await this.#getHttpApi()).fetch({
-                url: `${__dirname}/../../api/logout`,
-                method: METHOD_POST
-            });
-
-            return {
-                ok: true
-            };
-        } catch (error) {
-            console.error(error);
-
-            const localization_api = await this.#getLocalizationApi();
-
-            return {
-                ok: false,
-                "error-messages": [
-                    Object.fromEntries(await Promise.all(Object.entries((await localization_api.getLanguages()).all).map(async ([
-                        language
-                    ]) => [
-                            language,
-                            await localization_api.translate(
-                                error instanceof Response ? "Server error!" : "Network error!",
-                                null,
-                                null,
-                                language
-                            )
-                        ]
-                    )))
-                ]
-            };
-        }
-    }
-
-    /**
      * @param {boolean} previous_get_result
      * @returns {Promise<void>}
      */
@@ -888,7 +788,7 @@ export class StudisSelfserviceFrontendApi {
         const get_loading_element = await this.#getLoadingElement();
 
         if (!previous_get_result || this.#previous_get_result === null) {
-            this.#previous_get_result = await this.#get();
+            this.#previous_get_result = await (await this.#getRequestService()).get();
         }
 
         get_loading_element.remove();
@@ -900,7 +800,7 @@ export class StudisSelfserviceFrontendApi {
                 async post => {
                     const post_loading_element = await this.#getLoadingElement();
 
-                    const post_result = await this.#post(
+                    const post_result = await (await this.#getRequestService()).post(
                         post
                     );
 
@@ -919,7 +819,7 @@ export class StudisSelfserviceFrontendApi {
 
                     const back_loading_element = await this.#getLoadingElement();
 
-                    const back_result = await this.#back();
+                    const back_result = await (await this.#getRequestService()).back();
 
                     back_loading_element.remove();
 
@@ -953,7 +853,7 @@ export class StudisSelfserviceFrontendApi {
 
                 const logout_loading_element = await this.#getLoadingElement();
 
-                const logout_result = await this.#logout();
+                const logout_result = await (await this.#getRequestService()).logout();
 
                 logout_loading_element.remove();
 
@@ -971,41 +871,5 @@ export class StudisSelfserviceFrontendApi {
                 this.#next();
             } : null
         );
-    }
-
-    /**
-     * @param {Post} post
-     * @returns {Promise<PostResult>}
-     */
-    async #post(post) {
-        try {
-            return await (await this.#getHttpApi()).fetch({
-                url: `${__dirname}/../../api/post`,
-                method: METHOD_POST,
-                data: post
-            });
-        } catch (error) {
-            console.error(error);
-
-            const localization_api = await this.#getLocalizationApi();
-
-            return {
-                ok: false,
-                "error-messages": [
-                    Object.fromEntries(await Promise.all(Object.entries((await localization_api.getLanguages()).all).map(async ([
-                        language
-                    ]) => [
-                            language,
-                            await localization_api.translate(
-                                error instanceof Response ? "Server error!" : "Network error!",
-                                null,
-                                null,
-                                language
-                            )
-                        ]
-                    )))
-                ]
-            };
-        }
     }
 }