+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2023-01-20-1
+
+Changes:
+
+- Split to commands
+
 ## v2023-01-18-1
 
 Changes: